Show the available functions for the 4 classifications of data that are supported (numeric, date, binary and categorical)

Show the available functions for the 4 classifications of data that are supported (numeric, date, binary and categorical)

Usage

availableFunctions(variableType = NULL)

Arguments

variableType

A choice between: "numeric", "date", "binary" or "categorical".

Value

A tibble with the available functions for a certain variable classification (or all if NULL)

Examples

# \donttest{
library(PatientProfiles)

availableFunctions()
#> # A tibble: 102 × 4
#>    variable_type format_key applied_function                              result
#>    <chr>         <chr>      <chr>                                         <chr> 
#>  1 numeric       mean       mean of the variable of interest.             numer…
#>  2 numeric       sd         standard deviation of the variable of intere… numer…
#>  3 numeric       median     median of the variable of interest.           numer…
#>  4 numeric       q05        qualtile of 05% the variable of interest.     numer…
#>  5 numeric       q10        qualtile of 10% the variable of interest.     numer…
#>  6 numeric       q15        qualtile of 15% the variable of interest.     numer…
#>  7 numeric       q20        qualtile of 20% the variable of interest.     numer…
#>  8 numeric       q25        qualtile of 25% the variable of interest.     numer…
#>  9 numeric       q30        qualtile of 30% the variable of interest.     numer…
#> 10 numeric       q35        qualtile of 35% the variable of interest.     numer…
#> # ℹ 92 more rows
availableFunctions("numeric")
#> # A tibble: 25 × 3
#>    format_key applied_function                                result 
#>    <chr>      <chr>                                           <chr>  
#>  1 mean       mean of the variable of interest.               numeric
#>  2 sd         standard deviation of the variable of interest. numeric
#>  3 median     median of the variable of interest.             numeric
#>  4 q05        qualtile of 05% the variable of interest.       numeric
#>  5 q10        qualtile of 10% the variable of interest.       numeric
#>  6 q15        qualtile of 15% the variable of interest.       numeric
#>  7 q20        qualtile of 20% the variable of interest.       numeric
#>  8 q25        qualtile of 25% the variable of interest.       numeric
#>  9 q30        qualtile of 30% the variable of interest.       numeric
#> 10 q35        qualtile of 35% the variable of interest.       numeric
#> # ℹ 15 more rows
availableFunctions("date")
#> # A tibble: 24 × 3
#>    format_key applied_function                            result 
#>    <chr>      <chr>                                       <chr>  
#>  1 mean       mean of the date of interest.               date   
#>  2 sd         standard deviation of the date of interest. numeric
#>  3 median     median of the date of interest.             date   
#>  4 q05        qualtile of 05% the variable of interest.   date   
#>  5 q10        qualtile of 10% the variable of interest.   date   
#>  6 q15        qualtile of 15% the variable of interest.   date   
#>  7 q20        qualtile of 20% the variable of interest.   date   
#>  8 q25        qualtile of 25% the variable of interest.   date   
#>  9 q30        qualtile of 30% the variable of interest.   date   
#> 10 q35        qualtile of 35% the variable of interest.   date   
#> # ℹ 14 more rows
availableFunctions("binary")
#> # A tibble: 27 × 3
#>    format_key applied_function                                result    
#>    <chr>      <chr>                                           <chr>     
#>  1 count      number of times that each category is observed. numeric   
#>  2 percentage percentage of individuals with that category    percentage
#>  3 mean       mean of the variable of interest.               numeric   
#>  4 sd         standard deviation of the variable of interest. numeric   
#>  5 median     median of the variable of interest.             numeric   
#>  6 q05        qualtile of 05% the variable of interest.       numeric   
#>  7 q10        qualtile of 10% the variable of interest.       numeric   
#>  8 q15        qualtile of 15% the variable of interest.       numeric   
#>  9 q20        qualtile of 20% the variable of interest.       numeric   
#> 10 q25        qualtile of 25% the variable of interest.       numeric   
#> # ℹ 17 more rows
availableFunctions("categorical")
#> # A tibble: 26 × 3
#>    format_key          applied_function                                   result
#>    <chr>               <chr>                                              <chr> 
#>  1 count               number of times that each category is observed.    numer…
#>  2 percentage          percentage of records with that category           numer…
#>  3 distinct_categories number of distinct categories                      numer…
#>  4 mean                mean number of records per category                numer…
#>  5 sd                  standard deviation of the number of records per c… numer…
#>  6 median              median number of records per category              numer…
#>  7 q05                 qualtile of 05% the variable of interest.          numer…
#>  8 q10                 qualtile of 10% the variable of interest.          numer…
#>  9 q15                 qualtile of 15% the variable of interest.          numer…
#> 10 q20                 qualtile of 20% the variable of interest.          numer…
#> # ℹ 16 more rows
# }