Format a summarised_characteristics object into a visual table.

`r lifecycle::badge("deprecated")`

Usage

formatCharacteristics(
  result,
  type = "gt",
  splitStrata = TRUE,
  format = c(`N (%)` = "<count> (<percentage>%)", N = "<count>",
    "<median> [<q25> - <q75>]", "<mean> (<sd>)", range = "<min> to <max>"),
  cdmName = TRUE,
  cohortName = TRUE,
  style = "default",
  minCellCount = 5,
  .options = list()
)

Arguments

result

A summarised_characteristics object.

type

Type of desired formatted table, possibilities: "gt", "flextable", "tibble".

splitStrata

Whether or not to split the strata, default is True.

format

The columns that the user wishes to see for the formatted table.

cdmName

Whether or not to display the cdm name, default is TRUE.

cohortName

Weather or not to display the cohort name, default is TRUE.

style

The style of the table output.

minCellCount

Default is 5, meaning results that are more than 0 but less than 5 will not be reported.

.options

See optionsTableCharacteristics() for default values.

Value

A tibble with a tidy version of the summarised_characteristics object.

Examples

# \donttest{
library(PatientProfiles)

cdm <- mockPatientProfiles()

cdm$cohort1 |>
  summariseCharacteristics()
#> Warning: `summariseCharacteristics()` was deprecated in PatientProfiles 0.8.0.
#> ℹ Please use `CohortCharacteristics::summariseCharacteristics()` instead.
#> ℹ adding demographics columns
#> ℹ summarising data
#> ℹ The following estimates will be computed:
#> • variable_00002: count, percentage
#> • cohort_start_date: min, q05, q25, median, q75, q95, max
#> • cohort_end_date: min, q05, q25, median, q75, q95, max
#> • variable_00004: min, q05, q25, median, q75, q95, max, mean, sd
#> • variable_00005: min, q05, q25, median, q75, q95, max, mean, sd
#> • variable_00003: min, q05, q25, median, q75, q95, max, mean, sd
#> ! Table is collected to memory as not all requested estimates are supported on
#>   the database side
#> → Start summary of data, at 2024-04-16 10:59:47.324431
#> ✔ Summary finished, at 2024-04-16 10:59:47.43846
#> ✔ summariseCharacteristics finished!
#> # A tibble: 45 × 16
#>    result_id cdm_name result_type        package_name package_version group_name
#>        <int> <chr>    <chr>              <chr>        <chr>           <chr>     
#>  1         1 PP_MOCK  summarised_charac… PatientProf… 0.8.0           cohort_na…
#>  2         1 PP_MOCK  summarised_charac… PatientProf… 0.8.0           cohort_na…
#>  3         1 PP_MOCK  summarised_charac… PatientProf… 0.8.0           cohort_na…
#>  4         1 PP_MOCK  summarised_charac… PatientProf… 0.8.0           cohort_na…
#>  5         1 PP_MOCK  summarised_charac… PatientProf… 0.8.0           cohort_na…
#>  6         1 PP_MOCK  summarised_charac… PatientProf… 0.8.0           cohort_na…
#>  7         1 PP_MOCK  summarised_charac… PatientProf… 0.8.0           cohort_na…
#>  8         1 PP_MOCK  summarised_charac… PatientProf… 0.8.0           cohort_na…
#>  9         1 PP_MOCK  summarised_charac… PatientProf… 0.8.0           cohort_na…
#> 10         1 PP_MOCK  summarised_charac… PatientProf… 0.8.0           cohort_na…
#> # ℹ 35 more rows
#> # ℹ 10 more variables: group_level <chr>, strata_name <chr>,
#> #   strata_level <chr>, variable_name <chr>, variable_level <chr>,
#> #   estimate_name <chr>, estimate_type <chr>, estimate_value <chr>,
#> #   additional_name <chr>, additional_level <chr>

CDMConnector::cdmDisconnect(cdm = cdm)
# }