Plot cohort_overlap objects.

Plot cohort_overlap objects.

Usage

plotCohortOverlap(
  result,
  cohortNameReference = NULL,
  cohortNameComparator = NULL,
  strataName = NULL,
  strataLevel = NULL,
  cdmName = NULL,
  variableName = "number subjects",
  facetBy = NULL,
  overlapLabel = "{cohort_name_reference}; {cohort_name_comparator}",
  uniqueCombinations = TRUE
)

Arguments

result

A cohort_overlap object.

cohortNameReference

Names of the reference cohorts to include.

cohortNameComparator

Names of the comparator cohorts to include.

strataName

Names of the strata names to include.

strataLevel

Names of the strata levels to include.

cdmName

Names of the databases to include.

variableName

Names of the variable names to include.

facetBy

Names of columns in the cohort_overlap table for faceting the ggplot object.

overlapLabel

A glue expression to identify each plotted cohort overlap.

uniqueCombinations

If TRUE, only unique combinations of reference and comparator plots will be plotted.

Value

A ggplot2.

Examples

# \donttest{
# }