Compute the age of the individuals at a certain date

Compute the age of the individuals at a certain date

Usage

addAge(
  x,
  indexDate = "cohort_start_date",
  ageName = "age",
  ageGroup = NULL,
  ageMissingMonth = 1,
  ageMissingDay = 1,
  ageImposeMonth = FALSE,
  ageImposeDay = FALSE,
  missingAgeGroupValue = "None",
  name = NULL
)

Arguments

x

Table with individuals in the cdm.

indexDate

Variable in x that contains the date to compute the age.

ageName

Name of the new column that contains age.

ageGroup

List of age groups to be added.

ageMissingMonth

Month of the year assigned to individuals with missing month of birth. By default: 1.

ageMissingDay

day of the month assigned to individuals with missing day of birth. By default: 1.

ageImposeMonth

Whether the month of the date of birth will be considered as missing for all the individuals.

ageImposeDay

Whether the day of the date of birth will be considered as missing for all the individuals.

missingAgeGroupValue

Value to include if missing age.

name

Name of the new table, if NULL a temporary table is returned.

Value

tibble with the age column added.

Examples

# \donttest{
cdm <- mockPatientProfiles()
#> Note: method with signature ‘DBIConnection#Id’ chosen for function ‘dbExistsTable’,
#>  target signature ‘duckdb_connection#Id’.
#>  "duckdb_connection#ANY" would also be valid

cdm$cohort1 |>
  addAge()
#> # Source:   table<og_001_1722186208> [10 x 5]
#> # Database: DuckDB v1.0.0 [unknown@Linux 6.5.0-1024-azure:R 4.4.1/:memory:]
#>    cohort_definition_id subject_id cohort_start_date cohort_end_date   age
#>                   <int>      <int> <date>            <date>          <int>
#>  1                    2         10 1989-12-07        2035-03-09         34
#>  2                    1          6 2008-07-12        2024-07-01         29
#>  3                    3          9 2115-09-12        2123-05-08        153
#>  4                    2          5 2023-02-10        2080-05-12         99
#>  5                    2          3 2092-05-07        2106-12-13        123
#>  6                    1          4 1913-01-03        1951-07-29          8
#>  7                    2          1 2014-09-07        2022-03-05         12
#>  8                    1          7 2086-10-14        2104-02-24        109
#>  9                    2          8 1983-11-11        1984-08-31         81
#> 10                    3          2 2074-12-24        2119-12-12        101
mockDisconnect(cdm = cdm)
# }