Query to add the age of the individuals at a certain date

`r lifecycle::badge("experimental")` Same as `addAge()`, except query is not computed to a table.

Usage

addAgeQuery(
  x,
  indexDate = "cohort_start_date",
  ageName = "age",
  ageGroup = NULL,
  ageMissingMonth = 1,
  ageMissingDay = 1,
  ageImposeMonth = FALSE,
  ageImposeDay = FALSE,
  missingAgeGroupValue = "None"
)

Arguments

x

Table with individuals in the cdm.

indexDate

Variable in x that contains the date to compute the age.

ageName

Name of the new column that contains age.

ageGroup

List of age groups to be added.

ageMissingMonth

Month of the year assigned to individuals with missing month of birth. By default: 1.

ageMissingDay

day of the month assigned to individuals with missing day of birth. By default: 1.

ageImposeMonth

Whether the month of the date of birth will be considered as missing for all the individuals.

ageImposeDay

Whether the day of the date of birth will be considered as missing for all the individuals.

missingAgeGroupValue

Value to include if missing age.

Value

tibble with the age column added.

Examples

# \donttest{
cdm <- mockPatientProfiles()

cdm$cohort1 |>
  addAgeQuery()
#> # Source:   SQL [10 x 5]
#> # Database: DuckDB v1.0.0 [unknown@Linux 6.5.0-1025-azure:R 4.4.1/:memory:]
#>    cohort_definition_id subject_id cohort_start_date cohort_end_date   age
#>                   <int>      <int> <date>            <date>          <int>
#>  1                    2          8 2006-03-19        2118-10-13         57
#>  2                    3         10 1993-03-16        2019-03-30         80
#>  3                    1          5 2080-05-11        2119-04-28         65
#>  4                    2          9 1996-09-02        2012-02-09         82
#>  5                    3          7 2101-03-16        2144-05-07         92
#>  6                    3          6 2001-03-16        2003-09-11         91
#>  7                    2          3 2014-03-12        2014-10-20          4
#>  8                    2          4 1956-12-02        1973-01-15          6
#>  9                    3          1 2010-04-25        2085-07-31         51
#> 10                    1          2 1970-07-21        1972-02-27         26
mockDisconnect(cdm = cdm)
# }