It creates column to indicate the flag overlap information between a table and a concept

It creates column to indicate the flag overlap information between a table and a concept

Usage

addConceptIntersectFlag(
  x,
  conceptSet,
  indexDate = "cohort_start_date",
  censorDate = NULL,
  window = list(c(0, Inf)),
  targetStartDate = "event_start_date",
  targetEndDate = "event_end_date",
  nameStyle = "{concept_name}_{window_name}",
  name = NULL
)

Arguments

x

Table with individuals in the cdm.

conceptSet

Concept set list.

indexDate

Variable in x that contains the date to compute the intersection.

censorDate

whether to censor overlap events at a date column of x

window

window to consider events in.

targetStartDate

Event start date to use for the intersection.

targetEndDate

Event end date to use for the intersection.

nameStyle

naming of the added column or columns, should include required parameters.

name

Name of the new table, if NULL a temporary table is returned.

Value

table with added columns with overlap information

Examples

# \donttest{
library(PatientProfiles)
cdm <- mockPatientProfiles()
concept <- dplyr::tibble(
  concept_id = c(1125315),
  domain_id = "Drug",
  vocabulary_id = NA_character_,
  concept_class_id = "Ingredient",
  standard_concept = "S",
  concept_code = NA_character_,
  valid_start_date = as.Date("1900-01-01"),
  valid_end_date = as.Date("2099-01-01"),
  invalid_reason = NA_character_
) %>%
  dplyr::mutate(concept_name = paste0("concept: ", .data$concept_id))
cdm <- CDMConnector::insertTable(cdm, "concept", concept)
result <- cdm$cohort1 %>%
  addConceptIntersectFlag(
    conceptSet = list("acetaminophen" = 1125315)
  ) %>%
  dplyr::collect()
#> Warning: ! `codelist` contains numeric values, they are casted to integers.
mockDisconnect(cdm = cdm)
# }