Add a column with the individual birth date

Add a column with the individual birth date

Usage

addDateOfBirth(
  x,
  dateOfBirthName = "date_of_birth",
  missingDay = 1,
  missingMonth = 1,
  imposeDay = FALSE,
  imposeMonth = FALSE,
  name = NULL
)

Arguments

x

Table in the cdm that contains 'person_id' or 'subject_id'.

dateOfBirthName

Name of the column to be added with the date of birth.

missingDay

Day of the individuals with no or imposed day of birth.

missingMonth

Month of the individuals with no or imposed month of birth.

imposeDay

Whether to impose day of birth.

imposeMonth

Whether to impose month of birth.

name

Name of the new table, if NULL a temporary table is returned.

Value

The function returns the table x with an extra column that contains the date of birth.

Examples

# \donttest{
library(PatientProfiles)
cdm <- mockPatientProfiles()
cdm$cohort1 %>%
  addDateOfBirth()
#> # Source:   table<og_081_1722186261> [10 x 5]
#> # Database: DuckDB v1.0.0 [unknown@Linux 6.5.0-1024-azure:R 4.4.1/:memory:]
#>    cohort_definition_id subject_id cohort_start_date cohort_end_date
#>                   <int>      <int> <date>            <date>         
#>  1                    1          8 2004-04-01        2004-04-10     
#>  2                    3          4 2087-11-27        2090-06-02     
#>  3                    2          5 2002-12-27        2026-07-08     
#>  4                    1          2 2092-05-22        2106-06-11     
#>  5                    2          3 2045-06-12        2063-07-11     
#>  6                    3          9 1991-10-16        2118-04-11     
#>  7                    3          7 2068-04-12        2157-01-31     
#>  8                    1          6 2018-11-14        2021-06-05     
#>  9                    1         10 2114-08-21        2116-12-23     
#> 10                    3          1 2113-07-28        2133-07-20     
#> # ℹ 1 more variable: date_of_birth <date>
mockDisconnect(cdm = cdm)
# }