Compute demographic characteristics at a certain date

Compute demographic characteristics at a certain date

Usage

addDemographics(
  x,
  indexDate = "cohort_start_date",
  age = TRUE,
  ageName = "age",
  ageMissingMonth = 1,
  ageMissingDay = 1,
  ageImposeMonth = FALSE,
  ageImposeDay = FALSE,
  ageGroup = NULL,
  missingAgeGroupValue = "None",
  sex = TRUE,
  sexName = "sex",
  missingSexValue = "None",
  priorObservation = TRUE,
  priorObservationName = "prior_observation",
  priorObservationType = "days",
  futureObservation = TRUE,
  futureObservationName = "future_observation",
  futureObservationType = "days",
  dateOfBirth = FALSE,
  dateOfBirthName = "date_of_birth",
  name = NULL
)

Arguments

x

Table with individuals in the cdm.

indexDate

Variable in x that contains the date to compute the demographics characteristics.

age

TRUE or FALSE. If TRUE, age will be calculated relative to indexDate.

ageName

Age variable name.

ageMissingMonth

Month of the year assigned to individuals with missing month of birth.

ageMissingDay

day of the month assigned to individuals with missing day of birth.

ageImposeMonth

TRUE or FALSE. Whether the month of the date of birth will be considered as missing for all the individuals.

ageImposeDay

TRUE or FALSE. Whether the day of the date of birth will be considered as missing for all the individuals.

ageGroup

if not NULL, a list of ageGroup vectors.

missingAgeGroupValue

Value to include if missing age.

sex

TRUE or FALSE. If TRUE, sex will be identified.

sexName

Sex variable name.

missingSexValue

Value to include if missing sex.

priorObservation

TRUE or FALSE. If TRUE, days of between the start of the current observation period and the indexDate will be calculated.

priorObservationName

Prior observation variable name.

priorObservationType

Whether to return a "date" or the number of "days".

futureObservation

TRUE or FALSE. If TRUE, days between the indexDate and the end of the current observation period will be calculated.

futureObservationName

Future observation variable name.

futureObservationType

Whether to return a "date" or the number of "days".

dateOfBirth

TRUE or FALSE, if true the date of birth will be return.

dateOfBirthName

dateOfBirth column name.

name

Name of the new table, if NULL a temporary table is returned.

Value

cohort table with the added demographic information columns.

Examples

# \donttest{
library(PatientProfiles)
cdm <- mockPatientProfiles()
cdm$cohort1 %>%
  addDemographics()
#> # Source:   table<og_122_1726257652> [10 x 8]
#> # Database: DuckDB v1.0.0 [unknown@Linux 6.5.0-1025-azure:R 4.4.1/:memory:]
#>    cohort_definition_id subject_id cohort_start_date cohort_end_date   age sex  
#>                   <int>      <int> <date>            <date>          <int> <chr>
#>  1                    2          2 1936-02-11        1948-05-26         29 Male 
#>  2                    3          5 2005-05-07        2005-07-23         42 Male 
#>  3                    2          8 1977-03-24        1980-10-05         22 Male 
#>  4                    3          6 1927-08-15        1928-04-28          2 Fema…
#>  5                    1          4 1983-12-07        1989-03-15         37 Fema…
#>  6                    3         10 1939-11-12        1950-11-07         12 Fema…
#>  7                    2          3 1901-01-02        1905-04-21          0 Fema…
#>  8                    3          1 1956-09-28        1980-11-05         23 Male 
#>  9                    2          7 1982-01-30        1996-07-07          3 Male 
#> 10                    1          9 1957-07-05        1958-07-23         12 Fema…
#> # ℹ 2 more variables: prior_observation <int>, future_observation <int>
mockDisconnect(cdm = cdm)
# }