It adds all demographics columns to the given cohort table: Age, Sex, PriorHistory, and ageGroup if desired

It adds all demographics columns to the given cohort table: Age, Sex, PriorHistory, and ageGroup if desired

Usage

addDemographics(
  x,
  cdm,
  indexDate = "cohort_start_date",
  age = TRUE,
  ageName = "age",
  ageDefaultMonth = 1,
  ageDefaultDay = 1,
  ageImposeMonth = TRUE,
  ageImposeDay = TRUE,
  ageGroup = NULL,
  sex = TRUE,
  sexName = "sex",
  priorHistory = TRUE,
  priorHistoryName = "prior_history",
  futureObservation = TRUE,
  futureObservationName = "future_observation",
  tablePrefix = NULL
)

Arguments

x

cohort table in which to add follow up of individuals

cdm

cdm with the person and observation_period tables to get the info for the individuals in the cohort

indexDate

name of the column with the date at which consider demographic information

age

TRUE or FALSE. If TRUE, age will be calculated relative to indexDate

ageName

Age variable name

ageDefaultMonth

Month of the year assigned to individuals with missing month of birth.

ageDefaultDay

day of the month assigned to individuals with missing day of birth.

ageImposeMonth

TRUE or FALSE. Whether the month of the date of birth will be considered as missing for all the individuals.

ageImposeDay

TRUE or FALSE. Whether the day of the date of birth will be considered as missing for all the individuals.

ageGroup

if not NULL, a list of ageGroup vectors

sex

TRUE or FALSE. If TRUE, sex will be identified

sexName

Sex variable name

priorHistory

TRUE or FALSE. If TRUE, days of between the start of the current observation period and the indexDate will be calculated

priorHistoryName

Prior history variable name

futureObservation

TRUE or FALSE. If TRUE, days between the indexDate and the end of the current observation period will be calculated

futureObservationName

Future observation variable name

tablePrefix

The stem for the permanent tables that will be created. If NULL, temporary tables will be used throughout.

Value

cohort table with the added demographic information columns

Examples

# \donttest{
library(PatientProfiles)
cdm <- mockPatientProfiles()
cdm$cohort1 %>% addDemographics(cdm)
#> # Source:   table<dbplyr_004> [4 x 8]
#> # Database: DuckDB 0.7.1 [unknown@Linux 5.15.0-1035-azure:R 4.2.3/:memory:]
#>   cohort_definition_id subject_id cohort_start_date cohort_end_date   age sex  
#>                  <dbl>      <dbl> <date>            <date>          <dbl> <chr>
#> 1                    1          1 2020-01-01        2020-04-01         26 Male 
#> 2                    1          1 2020-06-01        2020-08-01         26 Male 
#> 3                    1          2 2020-01-02        2020-02-02         NA NA   
#> 4                    2          3 2020-01-01        2020-03-01         NA NA   
#> # ℹ 2 more variables: prior_history <dbl>, future_observation <dbl>
# }