Indicate if a certain record is within the observation period

Indicate if a certain record is within the observation period

Usage

addInObservation(
  x,
  indexDate = "cohort_start_date",
  window = c(0, 0),
  completeInterval = FALSE,
  nameStyle = "in_observation",
  name = NULL
)

Arguments

x

Table with individuals in the cdm.

indexDate

Variable in x that contains the date to compute the observation flag.

window

window to consider events of.

completeInterval

If the individuals are in observation for the full window.

nameStyle

Name of the new columns to create, it must contain "window_name" if multiple windows are provided.

name

Name of the new table, if NULL a temporary table is returned.

Value

cohort table with the added binary column assessing inObservation.

Examples

# \donttest{
cdm <- mockPatientProfiles()
cdm$cohort1 %>%
  addInObservation()
#> # Source:   table<og_114_1718094450> [10 x 5]
#> # Database: DuckDB v0.10.2 [unknown@Linux 6.5.0-1021-azure:R 4.4.0/:memory:]
#>    cohort_definition_id subject_id cohort_start_date cohort_end_date
#>                   <int>      <int> <date>            <date>         
#>  1                    3          4 1975-05-15        2026-06-10     
#>  2                    3         10 2131-04-13        2131-06-21     
#>  3                    3          6 2090-09-06        2091-03-28     
#>  4                    2          1 2050-04-09        2103-02-11     
#>  5                    3          5 2092-03-13        2119-12-17     
#>  6                    3          2 1991-11-12        2002-09-09     
#>  7                    2          9 2077-11-15        2110-05-04     
#>  8                    3          7 1965-08-09        2000-09-08     
#>  9                    3          8 2075-02-08        2094-09-30     
#> 10                    1          3 2065-05-06        2077-12-22     
#> # ℹ 1 more variable: in_observation <int>
mockDisconnect(cdm = cdm)
# }