Query to add the sex of the individuals

`r lifecycle::badge("experimental")` Same as `addSex()`, except query is not computed to a table.

Usage

addSexQuery(x, sexName = "sex", missingSexValue = "None")

Arguments

x

Table with individuals in the cdm.

sexName

name of the new column to be added.

missingSexValue

Value to include if missing sex.

Value

table x with the added column with sex information.

Examples

# \donttest{
cdm <- mockPatientProfiles()
#> ! cohort columns will be reordered to match the expected order:
#>   cohort_definition_id, subject_id, cohort_start_date, and cohort_end_date.
#> ! cohort columns will be reordered to match the expected order:
#>   cohort_definition_id, subject_id, cohort_start_date, and cohort_end_date.
cdm$cohort1 %>%
  addSexQuery()
#> # Source:   SQL [10 x 5]
#> # Database: DuckDB v1.0.0 [unknown@Linux 6.5.0-1025-azure:R 4.4.1/:memory:]
#>    cohort_definition_id subject_id cohort_start_date cohort_end_date sex   
#>                   <int>      <int> <date>            <date>          <chr> 
#>  1                    2          9 1992-06-10        2016-02-23      Female
#>  2                    3          8 2042-10-07        2076-04-16      Female
#>  3                    1          1 1948-01-29        1951-09-03      Female
#>  4                    3          4 1907-08-23        1939-04-18      Male  
#>  5                    1          5 2047-07-03        2108-08-20      Male  
#>  6                    1         10 1982-06-20        1998-03-10      Male  
#>  7                    1          2 1992-04-25        2028-03-13      Male  
#>  8                    2          7 2043-01-15        2064-09-09      Male  
#>  9                    3          3 2015-08-13        2033-12-09      Male  
#> 10                    2          6 1976-12-11        1983-03-22      Male  
mockDisconnect(cdm = cdm)
# }