Compute time to intersect with an omop table.

Compute time to intersect with an omop table.

Usage

addTableIntersectDays(
  x,
  tableName,
  indexDate = "cohort_start_date",
  censorDate = NULL,
  window = list(c(0, Inf)),
  targetDate = startDateColumn(tableName),
  order = "first",
  nameStyle = "{table_name}_{window_name}",
  name = NULL
)

Arguments

x

Table with individuals in the cdm.

tableName

Name of the table to intersect with. Options: visit_occurrence, condition_occurrence, drug_exposure, procedure_occurrence, device_exposure, measurement, observation, drug_era, condition_era, specimen, episode.

indexDate

Variable in x that contains the date to compute the intersection.

censorDate

whether to censor overlap events at a specific date or a column date of x.

window

window to consider events in.

targetDate

Target date in tableName.

order

which record is considered in case of multiple records (only required for date and days options).

nameStyle

naming of the added column or columns, should include required parameters.

name

Name of the new table, if NULL a temporary table is returned.

Value

table with added columns with intersect information.

Examples

# \donttest{
cdm <- mockPatientProfiles()

cdm$cohort1 %>%
  addTableIntersectDays(tableName = "visit_occurrence")
#> # Source:   table<og_130_1722186320> [10 x 5]
#> # Database: DuckDB v1.0.0 [unknown@Linux 6.5.0-1024-azure:R 4.4.1/:memory:]
#>    cohort_definition_id subject_id cohort_start_date cohort_end_date
#>                   <int>      <int> <date>            <date>         
#>  1                    1          7 1944-03-28        2017-07-14     
#>  2                    3          9 2053-05-25        2087-02-14     
#>  3                    1          2 1972-03-09        1994-12-24     
#>  4                    2          4 2118-03-03        2119-05-12     
#>  5                    3         10 1907-08-02        1921-01-14     
#>  6                    3          8 2123-06-10        2136-07-06     
#>  7                    2          1 2035-01-19        2037-10-12     
#>  8                    2          6 2078-11-28        2092-02-14     
#>  9                    3          5 1974-08-27        2059-01-25     
#> 10                    3          3 2107-05-30        2129-09-06     
#> # ℹ 1 more variable: visit_occurrence_0_to_inf <dbl>

mockDisconnect(cdm = cdm)
# }