It creates a mock database for testing PatientProfiles package

It creates a mock database for testing PatientProfiles package

Usage

mockPatientProfiles(
  con = NULL,
  writeSchema = NULL,
  numberIndividuals = 10,
  ...,
  seed = NULL
)

Arguments

con

A DBI connection to create the cdm mock object.

writeSchema

Name of an schema on the same connection with writing permisions.

numberIndividuals

Number of individuals to create in the cdm reference.

...

User self defined tables to put in cdm, it can input as many as the user want.

seed

A number to set the seed. If NULL seed is not used.

Value

A mock cdm_reference object created following user's specifications.

Examples

# \donttest{
library(PatientProfiles)
library(CDMConnector)

cdm <- mockPatientProfiles()

mockDisconnect(cdm = cdm)
# }